+++ avocado/utils/runtime.py
